requirements_dev.txt
lark-parser == 0.6.4
setuptools
tox
pytest
coverage
pytest-cov
bumpversion